feat(lib): rename framework to model and make it public
